# Avoid unnecessary allocations in clean_string & use cheaper methods in span construction