--- a/lib/ssf/ssf_methods.rb
+++ b/lib/ssf/ssf_methods.rb
@@ -73,8 +73,13 @@ module Ssf
     end
 
     def self.clean_string(str)
-      # assigning a non-utf8 string to a protobuf field will throw an exception
-      str.to_s.encode(Encoding::UTF_8, invalid: :replace, undef: :replace)
+      str = str.to_s unless str.is_a?(String)
+      if str.encoding != Encoding::UTF_8 || !str.valid_encoding?
+        # assigning a non-utf8 string to a protobuf field will throw an exception
+        str.to_s.encode(Encoding::UTF_8, invalid: :replace, undef: :replace)
+      else
+        str
+      end
     end
   end
 end