#### Summary
^

#### Motivation
pay-server spends ~0.8% of time in `SSF::BaseClient#start_span_from_context` and ~0.4% in `Ssf::SSFSpan.clean_string` in key methods per https://hubble.corp.stripe.com/queries/djudd/de43a824

Before:
```
[dev] main:0> s = 'foo_bar'; puts Benchmark.measure {1_000_000.times {clean_string(s)}}
  1.616000   0.000000   1.616000 (  1.617001)
[dev] main:0> puts Benchmark.measure {1_000_000.times {Time.now.to_f}}
  0.544000   0.004000   0.548000 (  0.547377)
[dev] main:0> x = 2**63-1; puts Benchmark.measure {1_000_000.times {SecureRandom.random_number(x)}}
  0.768000   1.068000   1.836000 (  1.836925)
```

After:
```
[dev] main:0> s = 'foo_bar'; puts Benchmark.measure {1_000_000.times {clean_string(s)}}
  0.124000   0.000000   0.124000 (  0.126248)
[dev] main:0> puts Benchmark.measure {1_000_000.times {Process.clock_gettime(Process::CLOCK_REALTIME)}}
  0.108000   0.000000   0.108000 (  0.107385)
[dev] main:0> x = 2**63-1; puts Benchmark.measure {1_000_000.times {Random.rand(x)}}
  0.268000   0.000000   0.268000 (  0.266199)
```

I can't think of a reason we'd need secure random numbers for span IDs.

#### Test plan
Existing tests (passing now - thanks @bobby-stripe !)

#### Rollout/monitoring/revert plan
Bump pay-server